pdf parse/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
tqdm>=4.66.0
